added drive and auto gyro turn
Added drive subsystem. Contains four TalonSRX objects called leftDriveMotor1 leftDriveMotor2 rightDriveMotor1 and rightDriveMotor2. Contains a gyro object. Contains methods to set each side independently called setLeftMotorValue and setRightMotorValue. Contains a method called tankDrive to set the robot to drive. Contains methods to get and reset the gyro value called resetGyro and getGyroValue. Created TankDriveCommand that reads from a joystick, joystick0, and operates a tank drive. Created a PID subsystem that reads from the drive gyro and turns an arbitrary number of degrees. Created a command to activate it. Set that to the a button. Made it run in autonomous via a command group.